rewrite game server manager
