+++ src/gameservers/index.ts
@@ -1,0 +1,10 @@
+import { localServerManager } from './managers/local.js'
+
+// Conditionally switch exports
+const activeManager = localServerManager
+
+// eslint-disable-next-line prettier/prettier
+const { getGameServers, getGameServer, addGameServer, removeGameServer } = activeManager
+export { getGameServers, getGameServer, addGameServer, removeGameServer }
+
+export { GameServer, CleanGameServer } from './gameserver.js'
